+++ src/vm/eventpipe.cpp
@@ -9,6 +9,7 @@
 #include "eventpipe.h"
 #include "eventpipebuffermanager.h"
 #include "eventpipeconfiguration.h"
+#include "eventpipeeventpayload.h"
 #include "eventpipesessionprovider.h"
 #include "eventpipeevent.h"
 #include "eventpipeeventsource.h"
@@ -39,146 +40,6 @@ extern "C" void InitProvidersAndEvents();
 #else
 void InitProvidersAndEvents();
 #endif
-
-EventPipeEventPayload::EventPipeEventPayload(EventData *pEventData, unsigned int eventDataCount)
-{
-    CONTRACTL
-    {
-        NOTHROW;
-        GC_NOTRIGGER;
-        MODE_ANY;
-    }
-    CONTRACTL_END;
-
-    m_pData = NULL;
-    m_pEventData = pEventData;
-    m_eventDataCount = eventDataCount;
-    m_allocatedData = false;
-
-    S_UINT32 tmp_size = S_UINT32(0);
-    for (unsigned int i = 0; i < m_eventDataCount; i++)
-    {
-        tmp_size += S_UINT32(m_pEventData[i].Size);
-    }
-
-    if (tmp_size.IsOverflow())
-    {
-        // If there is an overflow, drop the data and create an empty payload
-        m_pEventData = NULL;
-        m_eventDataCount = 0;
-        m_size = 0;
-    }
-    else
-    {
-        m_size = tmp_size.Value();
-    }
-}
-
-EventPipeEventPayload::~EventPipeEventPayload()
-{
-    CONTRACTL
-    {
-        NOTHROW;
-        GC_NOTRIGGER;
-        MODE_ANY;
-    }
-    CONTRACTL_END;
-
-    if (m_allocatedData && m_pData != NULL)
-    {
-        delete[] m_pData;
-        m_pData = NULL;
-    }
-}
-
-void EventPipeEventPayload::Flatten()
-{
-    CONTRACTL
-    {
-        NOTHROW;
-        GC_NOTRIGGER;
-        MODE_ANY;
-    }
-    CONTRACTL_END;
-
-    if (m_size > 0)
-    {
-        if (!IsFlattened())
-        {
-            BYTE *tmp_pData = new (nothrow) BYTE[m_size];
-            if (tmp_pData != NULL)
-            {
-                m_allocatedData = true;
-                CopyData(tmp_pData);
-                m_pData = tmp_pData;
-            }
-        }
-    }
-}
-
-void EventPipeEventPayload::CopyData(BYTE *pDst)
-{
-    CONTRACTL
-    {
-        NOTHROW;
-        GC_NOTRIGGER;
-        MODE_ANY;
-    }
-    CONTRACTL_END;
-
-    if (m_size > 0)
-    {
-        if (IsFlattened())
-        {
-            memcpy(pDst, m_pData, m_size);
-        }
-
-        else if (m_pEventData != NULL)
-        {
-            unsigned int offset = 0;
-            for (unsigned int i = 0; i < m_eventDataCount; i++)
-            {
-                memcpy(pDst + offset, (BYTE *)m_pEventData[i].Ptr, m_pEventData[i].Size);
-                offset += m_pEventData[i].Size;
-            }
-        }
-    }
-}
-
-BYTE *EventPipeEventPayload::GetFlatData()
-{
-    CONTRACTL
-    {
-        NOTHROW;
-        GC_NOTRIGGER;
-        MODE_ANY;
-    }
-    CONTRACTL_END;
-
-    if (!IsFlattened())
-    {
-        Flatten();
-    }
-    return m_pData;
-}
-
-void EventPipeProviderCallbackDataQueue::Enqueue(EventPipeProviderCallbackData *pEventPipeProviderCallbackData)
-{
-    SListElem<EventPipeProviderCallbackData> *listnode = new SListElem<EventPipeProviderCallbackData>(); // throws
-    listnode->m_Value = *pEventPipeProviderCallbackData;
-    list.InsertTail(listnode);
-}
-
-bool EventPipeProviderCallbackDataQueue::TryDequeue(EventPipeProviderCallbackData *pEventPipeProviderCallbackData)
-{
-    if (list.IsEmpty())
-        return false;
-
-    SListElem<EventPipeProviderCallbackData> *listnode = list.RemoveHead();
-    *pEventPipeProviderCallbackData = listnode->m_Value;
-    delete listnode;
-    return true;
-}
 
 void EventPipe::Initialize()
 {
