+++ src/vm/eventpipebuffer.cpp
@@ -5,6 +5,7 @@
 #include "common.h"
 #include "eventpipe.h"
 #include "eventpipeeventinstance.h"
+#include "eventpipeeventpayload.h"
 #include "eventpipebuffer.h"
 #include "eventpipebuffermanager.h"
 
@@ -89,7 +90,7 @@ bool EventPipeBuffer::WriteEvent(Thread *pThread, EventPipeSession &session, Eve
 
         EventPipeEventInstance *pInstance = new (m_pCurrent) EventPipeEventInstance(
             event,
-            (pThread == NULL) ? 
+            (pThread == NULL) ?
 #ifdef FEATURE_PAL
                 ::PAL_GetCurrentOSThreadId()
 #else
